Cleaner requires in shipit.rb
