+++ lib/shipit/null_serializer.rb
@@ -1,0 +1,13 @@
+module Shipit
+  module NullSerializer
+    extend self
+
+    def load(object)
+      object
+    end
+
+    def dump(object)
+      object
+    end
+  end
+end
