+++ test/jobs/github_sync_job_test.rb
@@ -29,7 +29,7 @@ class GithubSyncJobTest < ActiveSupport::TestCase
   test "#fetch_missing_commits returns the commits in the reverse order if it doesn't know the parent" do
     last = stub(sha: 123)
     first = stub(sha: 345)
-    FirstParentCommitsIterator.any_instance.stubs(:each).multiple_yields(last, first)
+    Shipit::FirstParentCommitsIterator.any_instance.stubs(:each).multiple_yields(last, first)
     @job.stubs(lookup_commit: nil)
 
     commits, parent = @job.fetch_missing_commits { stub }
@@ -40,7 +40,7 @@ class GithubSyncJobTest < ActiveSupport::TestCase
   test "#fetch_missing_commits returns the commits in the reverse order if it knows the parent" do
     last = stub(sha: 123)
     first = stub(sha: 345)
-    FirstParentCommitsIterator.any_instance.stubs(:each).multiple_yields(last, first)
+    Shipit::FirstParentCommitsIterator.any_instance.stubs(:each).multiple_yields(last, first)
     @job.stubs(:lookup_commit).with(123).returns(nil)
     @job.stubs(:lookup_commit).with(345).returns(first)
